requirements.txt
python-pptx
numpy
pandas
scikit-learn